gles2rice: Renamed OGLFragmentShaders => OGLES2FragmentShaders.
The fragment shader source code is so different between up and down-
stream that it might make more sense to maintain it as a completely
separate class.  The amount of code shared between the two is minimal,
and it's probably a lot easier to manually synchronize these small
sections than to maintain two very different implementations in the
same file.